Initial Upload
First repo for the Isometric Battle System I've been having an idea for EB, although the naming scheme for the code is uh.... questionable at best. The main code is in NPC.ccs, later I'll isolate the main part of the code into it's own file later, but now it's just for testing purposes. Right now I have Ness finished with actions he can take for his menu and status ailments. Need to do Paula, Jeff, and Poo's, but it should be a copy/paste effort to do from Ness so it should be easy to do??? Will definitely need the help of the community to test this out and see where it breaks so I can refine it better.